Try using `trim.sample` in `quantify.outliers`
Try using `trim.sample` in `quantify.outliers`, specifically in the
block calculating the trimmed mean z-scores.

A comparison of the output on a test sample suggests that there are
slight differences between the old and new versions for `trim` > 0,
suggesting that there's a slight difference in the indices being
chosen.