Define function `least.significant.digit`
Define function `least.significant.digit` in
`2.Distribution_Identification.R`.  This encapsulates code for
determining a 'minimum value' based on the least significant digit of
a vector of numbers.  This minimum value is used in a few places to
ensure that all values in a vector of numbers is larger than
zero.  (Similar code appears in `1.Outlier_Detection.R`, which should
also be refactored, but this can probably wait until code is migrated
to a dedicated package repo.)

Replace existing code defining `add.minimum.value` as a global
variable with the new function definition.