Def `identify.bic.optimal.distribution.residuals`
Define function `identify.bic.optimal.distribution.residuals` in
`2.Distribution_Identification.R` to determine the optimal
distribution of the residuals calculated earlier in the script.  This
function shares similarities with `identify.bic.optimal.distribution`
defined in `1.Outlier_Detection.R` in 13b1307.  In particular, like
that function, it returns the code corresponding to the optimal
distribution rather than a vector of the BIC values.