Refactor distribution identification script (#24)
* Define function `least.significant.digit`

Define function `least.significant.digit` in
`2.Distribution_Identification.R`.  This encapsulates code for
determining a 'minimum value' based on the least significant digit of
a vector of numbers.  This minimum value is used in a few places to
ensure that all values in a vector of numbers is larger than
zero.  (Similar code appears in `1.Outlier_Detection.R`, which should
also be refactored, but this can probably wait until code is migrated
to a dedicated package repo.)

Replace existing code defining `add.minimum.value` as a global
variable with the new function definition.

* Remove duplicate definition of `trim.sample`

Remove duplicate definition of `trim.sample` in
`2.Distribution_Identification.R`.  (We will need to rely on the
definition in `1.Outlier_Detection.R`, but as with the previous
commit, we'll not worry about the details that given that the code
will ultimately be migrated to a package-specific repo.)

* Define fun `calculate.residuals.observed.data`

Define function `calculate.residuals.observed.data` in
`2.Distribution_Identification.R` to calculate the residuals between
the quantiles of the observed data (for a single gene) and the
quantiles of the optimal distribution of the observed data (determined
in `1.Outlier_Detection.R`).

This encapsulates top-level code into a function and includes the
following minor changes:

- Include a parameter `distr` for the code corresponding to the
  optimal distribution for each gene.  This is as a result of changes
  in `1.Outlier_Detection.R` where the code was no longer appended to
  the data.
- Use function `least.significant.digit` to define the minimum value
  used to ensure that all values in the data are greater than zero.
- Don't sort the data.  This isn't necessary because the residuals are
  ultimately calculated based on the values of `ppoints()`, the return
  value of which is already sorted.
- Edit for style.

* Fix whitespace

* Replace code for calculating residuals

Replace code for calculating residuals with a call to
`calculate.residuals.observed.data` in
`2.Distribution_Identification.R`.

* Clean up code for calculating trimmed residuals

Clean up code for calculating trimmed residuals in
`2.Distribution_Identification.R`.  This includes using the updated
definition of the function `trim.sample`.

* Def `identify.bic.optimal.distribution.residuals`

Define function `identify.bic.optimal.distribution.residuals` in
`2.Distribution_Identification.R` to determine the optimal
distribution of the residuals calculated earlier in the script.  This
function shares similarities with `identify.bic.optimal.distribution`
defined in `1.Outlier_Detection.R` in 13b1307.  In particular, like
that function, it returns the code corresponding to the optimal
distribution rather than a vector of the BIC values.

* Reorganize `2.Distribution_Identification.R`

Move the definition of `identify.bic.optimal.distribution.residuals`
in `2.Distribution_Identification.R` with the other new function
definitions.

* Change residue -> residual

Change residue -> residual in a few more places in
`2.Distribution_Identification.R`.

* Replace code for determining optimal distribution

Replace code for determining the optimal distribution of residuals in
`2.Distribution_Identification.R` with a call to
`identify.bic.optimal.distribution.residuals`.

* Use function `least.significant.digit`

Replace manual calculations with calls to function
`least.significant.digit` in `1.Outlier_Detection.R`.

* Change residue -> residual

Change residue -> residual in `2.Distribution_Identification.R`.